+++ tests/manifests/test_manifest.py
@@ -1078,3 +1078,20 @@ def test_multi_nested_type_missmatch_with_partial(manifest_type, tmp_path, rc):
                   |   |   |   | meta.version.id                | integer |          | open   |
                   |   |   |   | meta                           | object  |          | open   |
             ''', manifest_type)
+
+
+@pytest.mark.manifests('internal_sql', 'csv')
+def test_enum_level(
+    manifest_type: str,
+    tmp_path,
+    rc
+):
+    check(tmp_path=tmp_path, rc=rc, table='''
+        d | r | b | m | property | type    | prepare | level | title
+        datasets/gov/example     |         |         |       |
+                                 |         |         |       |
+          |   |   | Data         |         |         |       |
+          |   |   |   | value    | integer |         |       |
+                                 | enum    | 1       | 3     | Positive
+                                 |         | 2       | 3     | Negative
+    ''', manifest_type=manifest_type)