+++ CHANGES.rst
@@ -30,11 +30,17 @@ Improvements:
 
   .. _#1011: https://github.com/atviriduomenys/spinta/issues/1011
 
+
+- added enum level support, allowing to indicate a level for enum. (`#982`_)
+
+  .. _#982: https://github.com/atviriduomenys/spinta/issues/982
+
 - Standardized `postgresql` naming convention, now all new constraints and indexes should follow same naming
   scheme (`P#153`).
 
 - `spinta migrate` now tries to rename constraints and indexes (if the name only changed) instead of dropping them and
   adding them with correct name (`P#153`).
+
 
 Bug fix: